move DeferredCache into its own module
